cleaning and renaming
