+++ src/main.cpp
@@ -6,10 +6,10 @@
 #include <model/model.h>
 #include <repository.h>
 #include <gui/gui_controller.h>
-#include "EspEEPROMRepository.h"
+#include "esp_eeprom_repository.h"
 #include "TFT_display_device.h"
 #include <light_sensor.h>
-#include "ArduinoConsole.h"
+#include "arduino_console.h"
 
 #define SCL_PIN 22
 #define SDA_PIN 21
@@ -27,7 +27,7 @@ const int READING_TIMEOUT = 20000;
 BH1750 device;
 LightSensor*  lightSensor;
 GuiController *guiController;
-//FlashMeterModel *model;
+
 Model* model;
 Repository* repository;
 ArduinoConsole console;
